Guardado, Validaciones, Mostrar
Generar capas restantes para hacer funcionar el proyecto
Hacer que se guarden los pacientes
Hacer que se vean los pacientes

+++ Fuentes/REPCovid/app/src/main/java/com/example/repcovid/dto/Paciente.java
@@ -10,7 +10,7 @@ public class Paciente implements Serializable {
     private String fechaExamen;
     private String areaTrabajo;
     private boolean sintomas;
-    private int temperatura;
+    private float temperatura;
     private boolean tos;
     private int presionArterial;
 
@@ -70,11 +70,11 @@ public class Paciente implements Serializable {
         this.sintomas = sintomas;
     }
 
-    public int getTemperatura() {
+    public float getTemperatura() {
         return temperatura;
     }
 
-    public void setTemperatura(int temperatura) {
+    public void setTemperatura(float temperatura) {
         this.temperatura = temperatura;
     }
